+++ src/media/genre_item.h
@@ -36,7 +36,7 @@ extern "C" {
 typedef struct genre_item {
     LIBRARY_ITEM_COMMON
 
-    unsigned int i_id;
+    int64_t i_id;
     char* psz_name;
 } genre_item;
